+++ tests/EscaladeTestCase.php
@@ -58,4 +58,81 @@ abstract class EscaladeTestCase extends DbTestCase
         Session::destroy();
         $_SESSION['glpi_currenttime'] = $ctime;
     }
+
+    public function climbWithTimelineButton(\Ticket $ticket, \Group $group, array $options): void
+    {
+        $options['ticket_details'] = array_merge(
+            $options['ticket_details'] ?? [],
+            [
+                'id' => $ticket->getID(),
+            ],
+        );
+        $_POST['comment'] = $options['comment'];
+        \PluginEscaladeTicket::timelineClimbAction($group->getID(), $ticket->getID(), $options);
+        $ticketgroup = new \Group_Ticket();
+        $is_escalate = $ticketgroup->getFromDBByCrit([
+            'tickets_id' => $ticket->getID(),
+            'groups_id'  => $group->getID(),
+        ]);
+        $this->assertTrue($is_escalate);
+        if (isset($options['is_observer_checkbox']) && $options['is_observer_checkbox']) {
+            $ticket_user = new \Ticket_User();
+            $is_observer = $ticket_user->getFromDBByCrit([
+                'type'       => \CommonITILActor::OBSERVER,
+                'tickets_id' => $ticket->getID(),
+                'users_id'   => Session::getLoginUserID(),
+            ]);
+            $this->assertTrue($is_observer);
+        }
+    }
+
+    public function climbWithHistoryButton(\Ticket $ticket, \Group $group): void
+    {
+        \PluginEscaladeTicket::climb_group($ticket->getID(), $group->getID(), true);
+        $ticketgroup = new \Group_Ticket();
+        $is_escalate = $ticketgroup->getFromDBByCrit([
+            'tickets_id' => $ticket->getID(),
+            'groups_id'  => $group->getID(),
+        ]);
+        $this->assertTrue($is_escalate);
+    }
+
+    public function climbWithSolvedTicket(\Ticket $ticket, \Group $group, array $solution_options = []): void
+    {
+        $this->createItem(\ITILSolution::class, array_merge([
+            'content' => 'Test Solution',
+            'itemtype' => $ticket->getType(),
+            'items_id' => $ticket->getID(),
+            'users_id' => Session::getLoginUserID(),
+        ], $solution_options));
+        $ticketgroup = new \Group_Ticket();
+        $is_escalate = $ticketgroup->getFromDBByCrit([
+            'tickets_id' => $ticket->getID(),
+            'groups_id'  => $group->getID(),
+        ]);
+        $this->assertTrue($is_escalate);
+    }
+
+    public function climbWithRejectSolutionTicket(\Ticket $ticket, \Group $group, array $followup_options = []): void
+    {
+        $_POST['add_reopen'] = 1;
+        $this->createItem(
+            \ITILFollowup::class,
+            array_merge([
+                'itemtype'   => 'Ticket',
+                'items_id'   => $ticket->getID(),
+                'add_reopen'   => '1',
+                'content'      => 'reopen followup',
+            ], $followup_options),
+            [
+                'add_reopen',
+            ],
+        );
+        $ticketgroup = new \Group_Ticket();
+        $is_escalate = $ticketgroup->getFromDBByCrit([
+            'tickets_id' => $ticket->getID(),
+            'groups_id'  => $group->getID(),
+        ]);
+        $this->assertTrue($is_escalate);
+    }
 }